Mention firmware override
I have debugged two separate issues caused by this functionality:
- A machine had test firmware added to it while debugging. This test
  firmware was never removed. Several software updates later, the
  application's messaging protocol was changed in a non-backwards
  compatible way, which resulted in strange breakage on this one
  machine. Fortunately the logs showed that the firmware was being
  overridden!
- An older firmware version was deployed to a new board that it wasn't
  intended to be used with. The result worked but the software update
  functionality was broken, and so it wasn't discovered until the next
  update. This lead to the technicians having to update the firmware to a
  non-broken version manually with a debugger.

I think it's generally useful, but somewhat dangerous. Ensuring that the
firmware is automatically phased out when the application is updated
would have solved the first issue. The second one is trickier, since it
was essentially done as a well meaning workaround for production.
Clearly communicating in the UI that this was for test purposes only
might have helped; a real solution would be that the firmware blob has
metadata about which boards it supports, but that's somewhat trickier
to pull off.